New mock_allocation_registry fixture. Fix mock_allocation_infile and broken tests

+++ tests/cli/conftest.py
@@ -20,14 +20,15 @@ import contextlib
 import json
 import os
 import shutil
-import sys
 
 import pytest
 from click.testing import CliRunner
 
+from nucypher.blockchain.eth.actors import ContractAdministrator
 from nucypher.blockchain.eth.registry import AllocationRegistry, InMemoryContractRegistry
 from nucypher.config.characters import UrsulaConfiguration
 from nucypher.utilities.sandbox.constants import (
+    MOCK_ALLOCATION_REGISTRY_FILEPATH,
     MOCK_CUSTOM_INSTALLATION_PATH,
     MOCK_ALLOCATION_INFILE,
     MOCK_REGISTRY_FILEPATH,
@@ -62,16 +63,31 @@ def nominal_federated_configuration_fields():
 def mock_allocation_infile(testerchain, token_economics):
     accounts = testerchain.unassigned_accounts
     allocation_data = [{'beneficiary_address': addr,
-                        'amount': token_economics.minimum_allowed_locked,
+                        'amount': 2 * token_economics.minimum_allowed_locked,
                         'duration_seconds': ONE_YEAR_IN_SECONDS}
                        for addr in accounts]
 
     with open(MOCK_ALLOCATION_INFILE, 'w') as file:
         file.write(json.dumps(allocation_data))
 
-    registry = AllocationRegistry(filepath=MOCK_ALLOCATION_INFILE)
-    yield registry
-    os.remove(MOCK_ALLOCATION_INFILE)
+    yield MOCK_ALLOCATION_INFILE
+    if os.path.isfile(MOCK_ALLOCATION_INFILE):
+        os.remove(MOCK_ALLOCATION_INFILE)
+
+
+@pytest.fixture(scope='module')
+def mock_allocation_registry(testerchain, test_registry, mock_allocation_infile):
+    admin = ContractAdministrator(registry=test_registry,
+                                  client_password=INSECURE_DEVELOPMENT_PASSWORD,
+                                  deployer_address=testerchain.etherbase_account)
+
+    admin.deploy_beneficiaries_from_file(allocation_data_filepath=mock_allocation_infile,
+                                         allocation_outfile=MOCK_ALLOCATION_REGISTRY_FILEPATH)
+
+    allocation_registry = AllocationRegistry(filepath=MOCK_ALLOCATION_REGISTRY_FILEPATH)
+    yield allocation_registry
+    if os.path.isfile(MOCK_ALLOCATION_REGISTRY_FILEPATH):
+        os.remove(MOCK_ALLOCATION_REGISTRY_FILEPATH)
 
 
 @pytest.fixture(scope='module', autouse=True)
